Add Maintainer Database and list view for it
This commit adds list of maintainers of a platform to page
/platforms/:id/maintainers.  This list is intended to serve as a
counterpart for http://maint.mandriva.com/listpkgs.php.

The commit adds a "view" on a build_list_packages table: it selects only
the latest builds for each package, and summarizes the information about
such packages in "maintainers" table.

This is a testing mode only, so the link to maintainers table is not
shown.